Refactor: Extract AnimatedChapterGridBackground and update version
This commit introduces the following changes:

- Extracted the `AnimatedChapterGridBackground` composable from `SagaReview.kt` into its own file under `app/src/main/java/com/ilustris/sagai/ui/animations/AnimatedChapterGridBackground.kt`. This improves code organization and reusability.
- Removed the `CharactersItem` composable from `SagaReview.kt` as it was no longer used.
- Updated the app version name from "1.1.2" to "1.2.0" in `app/build.gradle.kts`.
